+++ animation_nodes/utils/scene.py
@@ -1,0 +1,2 @@
+def getFPS(scene):
+    return scene.render.fps / scene.render.fps_base